gym-management-system/services/machines_services/machines_services/machines/views.py
# machines/views.py
from rest_framework import generics
from .models import Machine
from .serializers import MachineSerializer

class MachineListCreate(generics.ListCreateAPIView):
    queryset = Machine.objects.all()
    serializer_class = MachineSerializer

class MachineRetrieveUpdateDestroy(generics.RetrieveUpdateDestroyAPIView):
    queryset = Machine.objects.all()
    serializer_class = MachineSerializer