gym-management-system/services/machines_services/machines_services/requirements.txt
Django>=4.2
asgiref==3.8.1
djangorestframework==3.15.2
sqlparse==0.5.0
tzdata==2024.1
django-cors-headers==3.10.0 